Move Observer inside Evaluator
.. and add type synonyms in CallbackHandler
