+++ datasource/src/main/scala/quasar/plugin/googlebigtable/datasource/Evaluator.scala
@@ -22,8 +22,10 @@ import quasar.common.data.{CLong, CString, RObject, RValue}
 
 import scala.collection.JavaConverters._
 
-import cats.effect.ConcurrentEffect
+import cats.effect.{ConcurrentEffect, Sync}
 
+import com.google.api.gax.rpc.StreamController
+import com.google.api.gax.rpc.StateCheckingResponseObserver
 import com.google.cloud.bigtable.data.v2.BigtableDataClient
 import com.google.cloud.bigtable.data.v2.models.{Query => GQuery, Row, RowCell}
 
@@ -36,7 +38,6 @@ class Evaluator[F[_]: ConcurrentEffect](client: BigtableDataClient, query: GQuer
     val handler = Observer.handler[F](client.readRowsAsync(query, _))
     CallbackHandler.toStream[F, Row](handler, maxQueueSize).map(toRValue(_))
   }
-
 }
 
 object Evaluator {
@@ -62,5 +63,29 @@ object Evaluator {
       //"labels" -> RArray(rowCell.getLabels().asScala.map(CString(_)).toList),
       "timestamp" -> CLong(rowCell.getTimestamp()))
     (rowCell.getQualifier.toStringUtf8, rv)
+  }
+
+  class Observer[F[_]: Sync](callback: CallbackHandler.Callback[F, Row]) extends StateCheckingResponseObserver[Row] {
+
+    override protected def onStartImpl(controller: StreamController): Unit =
+      ()
+
+    override protected def onResponseImpl(row: Row): Unit =
+      callback(Right(Some(Sync[F].delay(row))))
+
+    override protected def onErrorImpl(t: Throwable): Unit =
+      callback(Left(t))
+
+    override protected def onCompleteImpl(): Unit =
+      callback(Right(None))
+
+  }
+
+  object Observer {
+    def handler[F[_]: Sync](f: Observer[F] => Unit): (Either[Throwable, Option[F[Row]]] => Unit) => F[Unit] = { cb =>
+      Sync[F].delay {
+        f(new Observer(cb))
+      }
+    }
   }
 }
